+++ test/store/database/test_database.py
@@ -96,6 +96,17 @@ def test_register_computations(
     assert deduplicated == expected_deduplication
 
 
+def test_read_computation_with_request_ts(backend_with_computation, default_computation_info):
+    computation_info = default_computation_info.model_copy()
+    computation_info.timestamp = datetime.datetime(2025, 1, 1, 12)
+    backend_with_computation.update_successful_computation(computation_info=computation_info)
+    db_computation_info = backend_with_computation.read_computation(
+        correlation_uuid=computation_info.correlation_uuid, state_actual_computation_time=True
+    )
+    assert db_computation_info.timestamp == datetime.datetime(2018, 1, 1, 12)
+    assert db_computation_info.message == 'The results were computed on the 2025-01-01 12:00:00'
+
+
 def test_read_duplicate_computation(
     default_plugin, default_backend_db, default_aoi_feature_geojson_pydantic, default_info, stop_time
 ):